Handle approveQuotes and move OfferState + some other things to their own files
